# app-AI/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MoneyMind API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserProfileLanguage =
  (typeof UserProfileLanguage)[keyof typeof UserProfileLanguage];

export const UserProfileLanguage = {
  en: "en",
  ar: "ar",
} as const;

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  monthlyIncome?: number;
  language: UserProfileLanguage;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type UpdateSettingsBodyLanguage =
  (typeof UpdateSettingsBodyLanguage)[keyof typeof UpdateSettingsBodyLanguage];

export const UpdateSettingsBodyLanguage = {
  en: "en",
  ar: "ar",
} as const;

export interface UpdateSettingsBody {
  monthlyIncome?: number;
  language?: UpdateSettingsBodyLanguage;
  name?: string;
}

export interface Expense {
  id: number;
  userId: number;
  amount: number;
  category: string;
  merchant?: string;
  description?: string;
  date: string;
  createdAt: string;
}

export interface CreateExpenseBody {
  amount: number;
  category: string;
  merchant?: string;
  description?: string;
  date?: string;
}

export type ExpenseSummaryByCategoryItem = {
  category: string;
  total: number;
  percentage: number;
};

export interface ExpenseSummary {
  totalSpent: number;
  byCategory: ExpenseSummaryByCategoryItem[];
  monthlyTrend: number;
  periodDays: number;
}

export type PortfolioHoldingType =
  (typeof PortfolioHoldingType)[keyof typeof PortfolioHoldingType];

export const PortfolioHoldingType = {
  stock: "stock",
  crypto: "crypto",
  etf: "etf",
} as const;

export interface PortfolioHolding {
  id: number;
  userId: number;
  symbol: string;
  name: string;
  type: PortfolioHoldingType;
  quantity: number;
  buyPrice: number;
  currentPrice: number;
  createdAt: string;
}

export type CreatePortfolioHoldingBodyType =
  (typeof CreatePortfolioHoldingBodyType)[keyof typeof CreatePortfolioHoldingBodyType];

export const CreatePortfolioHoldingBodyType = {
  stock: "stock",
  crypto: "crypto",
  etf: "etf",
} as const;

export interface CreatePortfolioHoldingBody {
  symbol: string;
  name: string;
  type: CreatePortfolioHoldingBodyType;
  quantity: number;
  buyPrice: number;
  currentPrice: number;
}

export interface UpdatePortfolioHoldingBody {
  currentPrice: number;
  quantity?: number;
}

export type PortfolioSummaryAllocationItem = {
  type: string;
  value: number;
  percentage: number;
};

export interface PortfolioSummary {
  totalValue: number;
  totalCost: number;
  totalPnl: number;
  totalPnlPercent: number;
  allocation: PortfolioSummaryAllocationItem[];
}

export type FinancialScoreLevel =
  (typeof FinancialScoreLevel)[keyof typeof FinancialScoreLevel];

export const FinancialScoreLevel = {
  Beginner: "Beginner",
  Saver: "Saver",
  Planner: "Planner",
  Investor: "Investor",
  Wealth_Builder: "Wealth Builder",
} as const;

export type FinancialScoreBreakdown = {
  spendingRatio: number;
  savingsRate: number;
  investmentRatio: number;
  consistency: number;
};

export interface FinancialScore {
  score: number;
  level: FinancialScoreLevel;
  breakdown: FinancialScoreBreakdown;
  previousScore?: number;
}

export interface Goal {
  id: number;
  userId: number;
  name: string;
  targetAmount: number;
  currentAmount: number;
  deadline?: string;
  createdAt: string;
}

export type BudgetPeriod = (typeof BudgetPeriod)[keyof typeof BudgetPeriod];

export const BudgetPeriod = {
  monthly: "monthly",
  weekly: "weekly",
} as const;

export interface Budget {
  id: number;
  userId: number;
  category: string;
  limitAmount: number;
  spentAmount: number;
  period: BudgetPeriod;
}

export interface GoalsAndBudgets {
  goals: Goal[];
  budgets: Budget[];
}

export interface CreateGoalBody {
  name: string;
  targetAmount: number;
  currentAmount?: number;
  deadline?: string;
}

export type CreateBudgetBodyPeriod =
  (typeof CreateBudgetBodyPeriod)[keyof typeof CreateBudgetBodyPeriod];

export const CreateBudgetBodyPeriod = {
  monthly: "monthly",
  weekly: "weekly",
} as const;

export interface CreateBudgetBody {
  category: string;
  limitAmount: number;
  period?: CreateBudgetBodyPeriod;
}

export type GamificationStatsAchievementsItem = {
  id: string;
  name: string;
  description: string;
  earned: boolean;
  earnedAt?: string;
};

export interface GamificationStats {
  points: number;
  streak: number;
  level: number;
  levelName: string;
  nextLevelPoints: number;
  achievements: GamificationStatsAchievementsItem[];
}

export type AwardPointsBodyAction =
  (typeof AwardPointsBodyAction)[keyof typeof AwardPointsBodyAction];

export const AwardPointsBodyAction = {
  add_expense: "add_expense",
  confirm_sms: "confirm_sms",
  chat_ai: "chat_ai",
  update_portfolio: "update_portfolio",
  complete_goal: "complete_goal",
  daily_login: "daily_login",
} as const;

export interface AwardPointsBody {
  action: AwardPointsBodyAction;
  points?: number;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  info: "info",
  warning: "warning",
  danger: "danger",
} as const;

export interface Alert {
  id: number;
  userId: number;
  title: string;
  message: string;
  severity: AlertSeverity;
  isRead: boolean;
  createdAt: string;
}

export interface ParseSmsBody {
  message: string;
}

export type ParsedSmsType = (typeof ParsedSmsType)[keyof typeof ParsedSmsType];

export const ParsedSmsType = {
  debit: "debit",
  credit: "credit",
} as const;

export interface ParsedSms {
  amount?: number;
  merchant?: string;
  type: ParsedSmsType;
  date?: string;
  category?: string;
  raw: string;
}

export interface DailyAction {
  action: string;
  reasoning: string;
  category?: string;
  targetAmount?: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type GetExpensesParams = {
  limit?: number;
  offset?: number;
  category?: string;
};