# app-AI/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MoneyMind API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gamification:
    get:
      operationId: getGamification
      tags: [gamification]
      summary: Get user's gamification stats (streak, points, level, achievements)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Gamification stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamificationStats"
  /gamification/award:
    post:
      operationId: awardPoints
      tags: [gamification]
      summary: Award points for completing an action
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwardPointsBody"
      responses:
        "200":
          description: Updated gamification stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamificationStats"
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: List user alerts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/read:
    put:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark an alert as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GamificationStats:
      type: object
      properties:
        points:
          type: integer
        streak:
          type: integer
        level:
          type: integer
        levelName:
          type: string
        nextLevelPoints:
          type: integer
        achievements:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              description:
                type: string
              earned:
                type: boolean
              earnedAt:
                type: string
                format: date-time
            required:
              - id
              - name
              - description
              - earned
      required:
        - points
        - streak
        - level
        - levelName
        - nextLevelPoints
        - achievements
    AwardPointsBody:
      type: object
      properties:
        action:
          type: string
          enum: [add_expense, confirm_sms, chat_ai, update_portfolio, complete_goal, daily_login]
        points:
          type: integer
      required:
        - action
    Alert:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, danger]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - message
        - severity
        - isRead
        - createdAt
...